Renaming internal `Promise` function to `makePromise`.
It wasn't really a constructor; it was a factory function. We do attach properties to `Promise.prototype` as if it were a constructor, but this seems to just be a convenience; we could just as well create a `promiseExemplar` object to replace `Promise.prototype` (now `makePromise.prototype`). Indeed, we already use `defer.prototype` while using lowercase to acknowledge that `defer` is not a constructor.