Fix long-standing bug in promiseSend protocol
I do not know whether this mistake ever manifested itself, but
promiseSend should never return a value, but previous attempts to to
remove the return broke several tests and I never bothered to isolate
the cause.  It turns out to be that "when" depended on the return value
in one case, in which it should not have.