Actually use jasmine-promise in Node!
It turns out all the asynchronous promise tests were just passing automatically. Now they are actually run.

This brought to light a small but mysterious issue with jasmine-promise, which had to be fixed.

In the process of debugging I strict-ified it because I wanted to make sure I wasn't doing something stupid. Let's keep that.