Don't mangle stack traces when an exception is seen twice
It's possible for the same exception to hit two different "end()"
points, causing it to pass through the stack trace formatter twice.

On the second pass, error.stack is already a string, and the formatter
mangles it by inserting newlines after every character.

The unit test included in this pull request depends on my earlier pull
request #94. That global error trap is pretty helpful for tests like
this.