Only shorten chain in valueOf if new value also isPromise.
This is the most conservative approach I could find to fix the test I
introduced in my previous commit.

Note that `valueOf` still returns the same value that it would have
returned in every case before this commit; it's just that `value` is held
back so that it can continue to behave as a promise in case a pending
nextTick callback tries to call `value.promiseDispatch`.