Make sure Q is in a real Node environment
This adds an additional check to ensure that the process global Q looks for is closer to the global that Nodejs environments expose. This is in light of updates to `node-process` a process shim used by Browserify:

https://github.com/defunctzombie/node-process/

This library no longer uses `setImmediate` if it is available, it uses `setTimeout` out of the box. Because Browserify is getting so popular as a moduling system, I think it is wise to add this check, so Browserify user's get all the extra `setImmediate` juice they possibly can.