Add Q.noConflict function
Fixes https://github.com/kriskowal/q/issues/677

This commit adds a function that is basically a noop. It simply alerts the user that they should not be using `Q.noConflict` when they are not using Q as a global.

If Q finds itself sourced as a `<script>` it will store the original value of `Q` and overwites `Q.noConflict` function with an actual implementation. `Q.noConflict` returns a refernce to `Q`.