Make npm test fail if jasmine-node throws
While working on writing some new (failing) tests,
`npm test` continued to report success.
If you had a syntax error in the spec file that made
jasmine-node unable to parse it, it would fail silently
and continue to `promises-aplus-tests`.

Adding `--captureExceptions` makes sure it exits when
an exception is thrown.

Also bumped the jasmine-node dependency to the latest version.